The main purpose of this commit is the ability to compile sfslite using the Clang/LLVM compiler. A lot of little things needed to be changed to get this to happen.
In addition, some actual bug fixes happened, most notably properly
generating the assembler instruction for reading the frame pointer on 64
bit systems.

Also, a fix for aios system, specifically aout. Make sure to flush out
anything that has been written in the destructor.

+++ arpc/axprt_pipe.C
@@ -234,7 +234,7 @@ axprt_pipe::output ()
   do {
     while (!syncpts.empty () && out->iovno () >= syncpts.front ())
       syncpts.pop_front ();
-    cnt = syncpts.empty () ? (size_t) -1
+    cnt = syncpts.empty () ? -1
       : int (syncpts.front () - out->iovno ());
   } while ((n = dowritev (cnt)) > 0);
